biweeklybudget/tests/acceptance/models/test_ofx_transaction.py
"""
################################################################################
Copyright 2016 Jason Antman <jason@jasonantman.com> <http://www.jasonantman.com>

    This file is part of biweeklybudget, also known as biweeklybudget.

    biweeklybudget is free software: you can redistribute it and/or modify
    it under the terms of the GNU Affero General Public License as published by
    the Free Software Foundation, either version 3 of the License, or
    (at your option) any later version.

    biweeklybudget is distributed in the hope that it will be useful,
    but WITHOUT ANY WARRANTY; without even the implied warranty of
    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
    GNU Affero General Public License for more details.

    You should have received a copy of the GNU Affero General Public License
    along with biweeklybudget.  If not, see <http://www.gnu.org/licenses/>.

The Copyright and Authors attributions contained herein may not be removed or
otherwise altered, except to add the Author attribution of a contributor to
this work. (Additional Terms pursuant to Section 7b of the AGPL v3)
################################################################################
While not legally required, I sincerely request that anyone who finds
bugs please submit them at <https://github.com/jantman/biweeklybudget> or
to me via email, and that you send any contributions or improvements
either as a pull request on GitHub, or to me via email.
################################################################################

AUTHORS:
Jason Antman <jason@jasonantman.com> <http://www.jasonantman.com>
################################################################################
"""

import pytest
from decimal import Decimal

from biweeklybudget.tests.acceptance_helpers import AcceptanceHelper
from biweeklybudget.models.ofx_transaction import OFXTransaction


@pytest.mark.acceptance
class TestFirstStatementByDate(AcceptanceHelper):

    def test_credit_one(self, testdb):
        trans = testdb.query(OFXTransaction).get((3, 'T2'))
        assert trans is not None
        assert trans.amount == Decimal('52.00')
        assert trans.statement_id == 6
        assert trans.statement.filename == '/stmt/CreditOne/0'
        res = trans.first_statement_by_date
        assert res is not None
        assert res.id == 4
        assert res.filename == '/stmt/CreditOne/3'
